change printing to file
